Reinstates symlinks to image READMEs
We have READMEs in the `image` folder so that they can be used as the
fallback 'main' document for a container. They are also helpful on Github.

These symlinks are required for the docs to build. Alternatives
to committing symlinks include creating them (or just copying the files)
in the `docs-build` package.json script. Feel free to do that if there
is a good reason not to use symlinks.